rename use-notes
